backend/requirements.txt
asgiref==3.6.0
Django==4.2.1
setuptools==67.3.2
sqlparse==0.4.3
tzdata==2022.7
wheel==0.38.4
djangorestframework~=3.14.0
django-cors-headers~=3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
PyJWT~=2.6.0
Pillow==9.4.0
gunicorn==20.1.0
psycopg2-binary==2.9.1